修复预计算 embedding 加载时的 dtype: bfloat16 转 float32 再转 numpy

+++ groot/vla/data/dataset/lerobot_sharded.py
@@ -1708,7 +1708,7 @@ class ShardedLeRobotSubLangSingleActionChunkDatasetDROIDPreprocessed(
                     if isinstance(text, list):
                         text = text[0]
                     emb = self._load_text_embedding(task_index)
-                    prompt_emb_by_text[text] = emb.numpy()
+                    prompt_emb_by_text[text] = emb.float().numpy()
             if prompt_emb_by_text:
                 step_data["prompt_emb_by_text"] = prompt_emb_by_text
